+++ collectors/etc/config.py
@@ -76,6 +76,7 @@ def get_defaults():
         'hosts': False,
         "monitoring_interface": None,
         "monitoring_port": 13280,
+        "namespace_prefix": "",
     }
 
     return defaults